Implement a second native extension for PyO3
[ci skip-build-wheels]
