+++ .github/workflows/test.yaml
@@ -102,6 +102,8 @@ jobs:
           '
         path: 'src/python/pants/engine/internals/native_engine.so
 
+          src/python/pants/engine/internals/native_engine_pyo3.so
+
           src/python/pants/engine/internals/native_engine.so.metadata'
     - name: Bootstrap Pants
       run: './pants --version
@@ -130,6 +132,8 @@ jobs:
       with:
         name: native_engine.so.${{ matrix.python-version }}.${{ runner.os }}
         path: 'src/python/pants/engine/internals/native_engine.so
+
+          src/python/pants/engine/internals/native_engine_pyo3.so
 
           src/python/pants/engine/internals/native_engine.so.metadata'
     - if: '!contains(env.COMMIT_MESSAGE, ''[ci skip-rust]'')'
@@ -241,6 +245,8 @@ jobs:
           '
         path: 'src/python/pants/engine/internals/native_engine.so
 
+          src/python/pants/engine/internals/native_engine_pyo3.so
+
           src/python/pants/engine/internals/native_engine.so.metadata'
     - name: Bootstrap Pants
       run: './pants --version
@@ -251,6 +257,8 @@ jobs:
       with:
         name: native_engine.so.${{ matrix.python-version }}.${{ runner.os }}
         path: 'src/python/pants/engine/internals/native_engine.so
+
+          src/python/pants/engine/internals/native_engine_pyo3.so
 
           src/python/pants/engine/internals/native_engine.so.metadata'
     - env:
