+++ build-support/bin/rust/bootstrap_code.sh
@@ -28,7 +28,9 @@ case "${KERNEL}" in
 esac
 
 readonly NATIVE_ENGINE_BINARY="native_engine.so"
+readonly NATIVE_ENGINE_BINARY_PYO3="native_engine_pyo3.so"
 readonly NATIVE_ENGINE_RESOURCE="${REPO_ROOT}/src/python/pants/engine/internals/${NATIVE_ENGINE_BINARY}"
+readonly NATIVE_ENGINE_RESOURCE_PYO3="${REPO_ROOT}/src/python/pants/engine/internals/${NATIVE_ENGINE_BINARY_PYO3}"
 readonly NATIVE_ENGINE_RESOURCE_METADATA="${NATIVE_ENGINE_RESOURCE}.metadata"
 
 function _build_native_code() {
@@ -37,15 +39,21 @@ function _build_native_code() {
   echo "${NATIVE_ROOT}/target/${MODE}/libengine.${LIB_EXTENSION}"
 }
 
+function _build_native_code_pyo3() {
+  # NB: See Cargo.toml with regard to the `extension-module` feature.
+  "${REPO_ROOT}/cargo" build --features=extension-module ${MODE_FLAG} -p engine_pyo3 || die
+  echo "${NATIVE_ROOT}/target/${MODE}/libengine_pyo3.${LIB_EXTENSION}"
+}
+
 function bootstrap_native_code() {
   # We expose a safety valve to skip compilation iff the user already has `native_engine.so`. This
   # can result in using a stale `native_engine.so`, but we trust that the user knows what
   # they're doing.
   if [[ "${SKIP_NATIVE_ENGINE_SO_BOOTSTRAP}" == "true" ]]; then
-    if [[ ! -f "${NATIVE_ENGINE_RESOURCE}" ]]; then
-      die "You requested to override bootstrapping native_engine.so via the env var" \
-        "SKIP_NATIVE_ENGINE_SO_BOOTSTRAP, but the file does not exist at" \
-        "${NATIVE_ENGINE_RESOURCE}. This is not safe to do."
+    if [[ ! -f "${NATIVE_ENGINE_RESOURCE}" || ! -f "${NATIVE_ENGINE_RESOURCE_PYO3}" ]]; then
+      die "You requested to override bootstrapping native_engine.so and native_engine_pyo3.so via the env var" \
+        "SKIP_NATIVE_ENGINE_SO_BOOTSTRAP, but the files do not exist at" \
+        "${NATIVE_ENGINE_RESOURCE} and ${NATIVE_ENGINE_BINARY_PYO3}. This is not safe to do."
     fi
     return
   fi
@@ -56,14 +64,18 @@ function bootstrap_native_code() {
   if [[ -f "${NATIVE_ENGINE_RESOURCE_METADATA}" ]]; then
     engine_version_in_metadata="$(sed -n 's/^engine_version: //p' "${NATIVE_ENGINE_RESOURCE_METADATA}")"
   fi
-  if [[ ! -f "${NATIVE_ENGINE_RESOURCE}" || "${engine_version_calculated}" != "${engine_version_in_metadata}" ]]; then
+  if [[ ! -f "${NATIVE_ENGINE_RESOURCE}" || ! -f "${NATIVE_ENGINE_RESOURCE_PYO3}" || "${engine_version_calculated}" != "${engine_version_in_metadata}" ]]; then
     echo "Building native engine"
     local -r native_binary="$(_build_native_code)"
+    local -r native_binary_pyo3="$(_build_native_code_pyo3)"
 
     # If bootstrapping the native engine fails, don't attempt to run pants
     # afterwards.
     if [[ ! -f "${native_binary}" ]]; then
-      die "Failed to build native engine."
+      die "Failed to build native engine, file missing at ${native_binary}."
+    fi
+    if [[ ! -f "${native_binary_pyo3}" ]]; then
+      die "Failed to build native engine, file missing at ${native_binary_pyo3}."
     fi
 
     # Pick up Cargo.lock changes if any caused by the `cargo build`.
@@ -72,8 +84,9 @@ function bootstrap_native_code() {
     # Create the native engine resource.
     # NB: On Mac Silicon, for some reason, first removing the old native_engine.so is necessary to avoid the Pants
     #  process from being killed when recompiling.
-    rm -f "${NATIVE_ENGINE_RESOURCE}"
+    rm -f "${NATIVE_ENGINE_RESOURCE}" "${NATIVE_ENGINE_RESOURCE_PYO3}"
     cp "${native_binary}" "${NATIVE_ENGINE_RESOURCE}"
+    cp "${native_binary_pyo3}" "${NATIVE_ENGINE_RESOURCE_PYO3}"
 
     # Create the accompanying metadata file.
     local -r metadata_file=$(mktemp -t pants.native_engine.metadata.XXXXXX)
