+++ cargo
@@ -9,6 +9,7 @@ source "${REPO_ROOT}/build-support/common.sh"
 
 PY="$(determine_python)"
 export PY
+export PYO3_PYTHON="${PY}"
 export PYTHON_SYS_EXECUTABLE="${PY}" # Consumed by the cpython crate.
 
 if ! command -v rustup &> /dev/null; then
