+++ src/python/pants/engine/internals/.gitignore
@@ -1,2 +1,3 @@
 /native_engine.so
+/native_engine_pyo3.so
 /native_engine.so.metadata
