+++ src/python/pants/engine/internals/native_engine_pyo3.pyi
@@ -1,0 +1,15 @@
+# Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+class PyExecutor:
+    def __init__(self, core_threads: int, max_threads: int) -> None: ...
+
+class PyStubCASBuilder:
+    def always_errors(self) -> PyStubCASBuilder: ...
+    def build(self, executor: PyExecutor) -> PyStubCAS: ...
+
+class PyStubCAS:
+    @classmethod
+    def builder(cls) -> PyStubCASBuilder: ...
+    @property
+    def address(self) -> str: ...
