+++ src/rust/engine/Cargo.toml
@@ -23,6 +23,7 @@ members = [
   "async_semaphore",
   "async_value",
   "concrete_time",
+  "engine_pyo3",
   "fs",
   "fs/brfs",
   "fs/fs_util",
@@ -61,6 +62,7 @@ default-members = [
   "async_semaphore",
   "async_value",
   "concrete_time",
+  "engine_pyo3",
   "fs",
   "fs/fs_util",
   "fs/store",
@@ -120,7 +122,6 @@ logging = { path = "logging" }
 nailgun = { path = "nailgun" }
 num_enum = "0.4"
 parking_lot = "0.11"
-pyo3 = { version = "0.13", features = ["extension-module"] }
 process_execution = { path = "process_execution" }
 rand = "0.8"
 regex = "1"
