+++ tests/python/pants_test/integration/remote_cache_integration_test.py
@@ -1,14 +1,15 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.engine.internals.native_engine import PyStubCAS
+from pants.engine.internals.native_engine_pyo3 import PyExecutor, PyStubCAS
 from pants.option.global_options import RemoteCacheWarningsBehavior
 from pants.option.scope import GLOBAL_SCOPE_CONFIG_SECTION
 from pants.testutil.pants_integration_test import run_pants
 
 
 def test_warns_on_remote_cache_errors():
-    cas = PyStubCAS.builder().always_errors().build()
+    executor = PyExecutor(core_threads=2, max_threads=4)
+    cas = PyStubCAS.builder().always_errors().build(executor)
 
     def run(behavior: RemoteCacheWarningsBehavior) -> str:
         pants_run = run_pants(
